pyproject.toml
[project]
name = "nonebot-plugin-dice-helper"
version = "0.1.2"
description = "NoneBot 可自定义骰子的骰子插件"
authors = [
    { name = "Demo" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "nonebot2>=2.4.4",
    "nonebot_adapter_onebot>=2.4.6",
    "nonebot-plugin-localstore>=0.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/CaptainDemo/nonebot-plugin-dice-helper"
Repository = "https://github.com/CaptainDemo/nonebot-plugin-dice-helper"

[tool.nonebot]
plugins = ["nonebot_plugin_dice_helper"]

[build-system]
requires = ["setuptools>=58.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

# =====================
# pytest 配置
# =====================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# =====================
# 覆盖率配置
# =====================

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"
